requirements.txt
autopep8==1.5.7
pycodestyle==2.7.0
python-dotenv==0.18.0
selenium==3.141.0
toml==0.10.2
urllib3==1.26.6
